IMPALA-10525: Add param to BuiltinsDb to defer initialization
BuiltinsDb currently initializes all the builtin functions on
initialization. Part of the initialization task is to interact with
the C++ code to fetch the signatures of the functions.  This doesn't
work if a third party extension wants to use the BuiltinsDb but does
not have access to the C++ library at runtime.

The solution is to add an alternative way to initalize the BuiltinsDb,
through a Loader class.

Change-Id: I1941a2494efe15b63514d849873eeb1c4ed8a981
Reviewed-on: http://gerrit.cloudera.org:8080/17093
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>