IMPALA-10531: Fix TmpFileMgrTest.TestCompressBufferManagementEncryptedRemoteUpload failed in exhaustive release build
Fixed TmpFileMgrTest testcases failed in exhaustive release build.
The cause is that in TmpFileGroup::AllocateRemoteSpace(), it uses
DCHECK to check the return of TmpFileRemote::AllocateSpace(), because
it always returns true, in the release build, the logic is optimized,
so TmpFileRemote::AllocateSpace() isn't called in the release build.
The solution is to use if instead of DCHECK.

Tests:
Reran TmpFileMgrTest/DiskIoMgrTest/BufferPoolTest in the release build.

Change-Id: I7320ae41957c44151b7ec17886c383e72c2546e4
Reviewed-on: http://gerrit.cloudera.org:8080/17101
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>