IMPALA-10518: Add ImpalaServer interface to retrieve executor membership.
This patch adds an interface to ImpalaServer to retrieve the current
executor membership snapshot from impalad for use by an external
frontend. This involves sending a thrift request to impalad and
receiving a thrift response. Refactored some code in exec-env into
a separate function in the impala namespace which makes it easier to
populate the needed information for an external frontend.

Testing:
 - Ran selected tests for sanity check (no impact is expected
   since this is adding a new interface):
    - Frontend tests (PlannerTest, CardinalityTest)
    - Backend tests under custom_cluster/test_executor_groups.py
 - Manually tested with external frontend to ensure it gets
   the executor membership snapshot.

Change-Id: Ie89b71f4555c368869ee7b9d6341756c60af12b5
Reviewed-on: http://gerrit.cloudera.org:8080/17181
Reviewed-by: Thomas Tauber-Marshall <tmarshall@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>