IMPALA-10593: Conditionally skip runtime filter for outer joins
Currently there is code that asserts that an Expr is not constant after
substituting SlotRefs with constant nulls.

For External FE, this restriction to be weakened.  In a case where
an Expr is checked and the Expr is not constant even after substituting
nulls, the result will be to not generate a runtime filter for that Expr.

Testing:

Manually tested with this query in the External FE:

select id, int_col, year, month from alltypessmall s
where s.int_col = (select count(*) from alltypestiny t where s.id = t.id)
order by id

Change-Id: I46462e2030731d97c4c88e364148c0093c025ab3
Reviewed-on: http://gerrit.cloudera.org:8080/17200
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>