IMPALA-10591: Handle failed ReleaseQueryBackends rpcs
When the admission control service is in use, coordinators will
retry failed ReleaseQueryBackends rpcs 3 times before giving up.
This can potentially result in resources not being released when they
are no longer in use.

This patch fixes the issue by automatically releasing all remaining
backends when ReleaseQuery is called in the context of the admission
control service (either as the result of a ReleaseQUery rpc or when
cleaning up a query's resources with the heartbeat mechanism).

Testing:
- Added a custom cluster test that simulates failed
  ReleaseQueryBackends rpcs and ensures that query resources are
  eventually released.

Change-Id: I22842b2fb8ee170b5e91f12cd83f57a5f5502ae9
Reviewed-on: http://gerrit.cloudera.org:8080/17208
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Thomas Tauber-Marshall <tmarshall@cloudera.com>