IMPALA-10605: Deflake test_refresh_native
This test uses a regex to parse the output of
describe database and extract the db properties. The regex assumes that there
will only be 1 key value pair which is broken when events processor is
running. The fix is to modify the regex so that it only extracts the
relevant function name prefix and its value.

Testing:
1. The test fails when events processor is enabled. After the patch
the test works as expected.

Change-Id: I1df35b9c5f2b21cc7172f03ff8611d46070d64c2
Reviewed-on: http://gerrit.cloudera.org:8080/17227
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>