IMPALA-10619: Minor refactoring of analytic function methods
The FIRST_VALUE, LAST_VALUE functions go through standardization
process in AnalyticExpr where they may be rewritten with different
number of parameters or with different window frame. In order for an
external FE to leverage this standardization, this patch creates a
wrapper method for FunctionCallExpr creation and does minor refactoring.

Also added accessor methods to AnalyticEvalNode and changed visibility
of couple of methods in PlanNode for use by external FE.

Testing: Ran PlannerTests. No new tests are added since this does not
change the existing behavior.

Change-Id: I39e4268c0c5500f09acf98357a80763c28f615c2
Reviewed-on: http://gerrit.cloudera.org:8080/17237
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Aman Sinha <amsinha@cloudera.com>
Reviewed-by: Aman Sinha <amsinha@cloudera.com>