IMPALA-10596: De-flake teardown in TestAdmissionControllerStress
Currently if the threads running queries in
TestAdmissionControllerStress hit an error, they close their client
which ultimately closes the query that it was running. If teardown()
runs after the client is closed, it tries to cancel the query that
the thread was running and hits an exception trying to cancel an
already closed query. This results in the pytest throws the exception
encountered in teardown() instead of the original exception that
caused the test to fail in the first place. This patch fixes this by
removing the query handle from the thread if the client is closed.

Testing:
Simulated hitting an error condition in the main thread that
initially triggered this condition.

Change-Id: I8aa8315d9f598ba80d13cd2091e3cc743c64ba77
Reviewed-on: http://gerrit.cloudera.org:8080/17256
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>