IMPALA-10637: Fixes bug in ValidWriteIdList comparison
For a transactional table, catalogd compares previous and current ValidWriteList
to determine more recent version out of the two and reloads table cache accordingly.
Because of a bug in ValidWriteIdList comparison, catalogD was not refreshing table 
metadata in the cache with more recent changes. As a result of which we were seeing 
inconsistencies in read after write into the table.

Tested by
  1. Adding a unit test to compare WriteIDLists.

Change-Id: Idaa4bcdbda1757a6451122efc505d1d483c879cc
Reviewed-on: http://gerrit.cloudera.org:8080/17276
Reviewed-by: Sourabh Goyal <sourabhg@cloudera.com>
Reviewed-by: Vihang Karajgaonkar <vihang@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>