IMPALA-10643: Allow the inclusion of jetty-client
We excluded all artifacts under the group of org.eclipse.jetty for
ranger-plugins-audit when bumping up CDP_BUILD_NUMBER to 4493826, which
is too stringent in that the artifact of jetty-client is actually
required by ranger-plugins-audit. Excluding jetty-client would thus
result in a NoClassDefFoundError at runtime. This patch removes the
block that excluded all artifacts of org.eclipse.jetty when the
dependency of ranger-plugins-audit is added, which allows
ranger-plugins-audit to pull in jetty-client, which in turn pulls in
jetty-http and jetty-io. In this regard, we also add these three
artifacts as allowed dependencies because all artifacts under
org.eclipse.jetty are banned in the section of bannedDependencies.

Testing:
 - Verified in a local development environment that Impala could build
   and that jetty-client-9.4.31.v20200723.jar is indeed on the class
   path in fe/target/build-classpath.txt.

Change-Id: I9087b7e6866f1500c66f42a74b3f8619e82c3bda
Reviewed-on: http://gerrit.cloudera.org:8080/17287
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>